requirements.txt
numpy
scikit-image
matplotlib
Pillow
scikit-learn
kneed
colormath
colorthief
palettable